Refactor
pkg/registry/* -> pkg/adapter/*
