+++ adapter/denyChecker/BUILD
@@ -11,7 +11,6 @@ go_library(
     deps = [
         "//adapter/denyChecker/config:go_default_library",
         "//pkg/adapter:go_default_library",
-        "//pkg/registry:go_default_library",
         "@com_github_google_go_genproto//googleapis/rpc/code:go_default_library",
         "@com_github_google_go_genproto//googleapis/rpc/status:go_default_library",
     ],
