+++ adapter/genericListChecker/adapter.go
@@ -18,11 +18,10 @@ package genericListChecker
 import (
 	"istio.io/mixer/adapter/genericListChecker/config"
 	"istio.io/mixer/pkg/adapter"
-	"istio.io/mixer/pkg/registry"
 )
 
 // Register records the existence of this adapter
-func Register(r registry.Registrar) error {
+func Register(r adapter.Registrar) error {
 	return r.RegisterCheckList(newAdapter())
 }
