+++ adapter/stdioLogger/BUILD
@@ -10,7 +10,6 @@ go_library(
     deps = [
         "//adapter/stdioLogger/config:go_default_library",
         "//pkg/adapter:go_default_library",
-        "//pkg/registry:go_default_library",
         "@com_github_hashicorp_go_multierror//:go_default_library",
     ],
 )
