+++ adapter/stdioLogger/stdioLogger.go
@@ -23,7 +23,6 @@ import (
 
 	"istio.io/mixer/adapter/stdioLogger/config"
 	"istio.io/mixer/pkg/adapter"
-	"istio.io/mixer/pkg/registry"
 
 	me "github.com/hashicorp/go-multierror"
 )
@@ -37,7 +36,7 @@ type (
 
 // Register adds the stdioLogger adapter to the list of logger.Aspects known to
 // mixer.
-func Register(r registry.Registrar) error { return r.RegisterLogger(&adapterState{}) }
+func Register(r adapter.Registrar) error { return r.RegisterLogger(&adapterState{}) }
 
 func (a *adapterState) Name() string { return "istio/stdioLogger" }
 func (a *adapterState) Description() string {
