+++ pkg/api/BUILD
@@ -9,11 +9,11 @@ go_library(
         "methodHandlers.go",
     ],
     deps = [
+        "//pkg/adapter:go_default_library",
         "//pkg/adapterManager:go_default_library",
         "//pkg/aspect:go_default_library",
         "//pkg/attribute:go_default_library",
         "//pkg/expr:go_default_library",
-        "//pkg/registry:go_default_library",
         "//pkg/tracing:go_default_library",
         "@com_github_golang_glog//:go_default_library",
         "@com_github_golang_protobuf//proto:go_default_library",
