--- pkg/registry/BUILD
@@ -1,18 +1,0 @@
-package(default_visibility = ["//visibility:public"])
-
-licenses(["notice"])
-
-load("@io_bazel_rules_go//go:def.bzl", "go_prefix", "go_library")
-
-go_prefix("istio.io/mixer/pkg/registry")
-
-go_library(
-    name = "go_default_library",
-    srcs = glob(
-        include = ["*.go"],
-        exclude = ["test_*.go"],
-    ),
-    deps = [
-        "//pkg/adapter:go_default_library",
-    ],
-)
